feat: refactored, self review
