+++ src/components/DashboardMetricChart/DashboardMetricChart.js
@@ -11,10 +11,10 @@ import SharedAxisToggle from '../../ducks/Studio/Chart/SharedAxisToggle'
 import { DesktopOnly } from '../Responsive'
 import Chart from '../../ducks/Chart'
 import { mapSizesToProps } from '../../utils/withSizes'
-import styles from './DashboardMetricChart.module.scss'
 import { updateTooltipSettings } from '../../ducks/dataHub/tooltipSettings'
 import { useChartColors } from '../../ducks/Chart/colors'
-import { CRYPTO_ERA_START_DATE } from '../../utils/dates'
+import { DEFAULT_INTERVAL_SELECTORS, INTERVAL_30_DAYS } from './utils'
+import styles from './DashboardMetricChart.module.scss'
 
 const CHART_HEIGHT = 400
 const CHART_PADDING_DESKTOP = {
@@ -29,47 +29,6 @@ const CHART_PADDING_MOBILE = {
   bottom: 0,
   left: 0
 }
-
-export const makeIntervalSelectors = ({
-  val,
-  label,
-  interval = '1d',
-  from
-}) => ({
-  value: val,
-  label: label,
-  requestParams: {
-    from: from || `utc_now-${val}`,
-    to: `utc_now`,
-    interval
-  }
-})
-
-export const makeMetric = (key, label) => {
-  return {
-    key,
-    label
-  }
-}
-
-export const INTERVAL_30_DAYS = makeIntervalSelectors({
-  val: '30d',
-  label: '1M',
-  interval: '3h'
-})
-
-export const DEFAULT_INTERVAL_SELECTORS = [
-  makeIntervalSelectors({ val: '1d', label: '1D', interval: '15m' }),
-  makeIntervalSelectors({ val: '1w', label: '1W', interval: '1h' }),
-  INTERVAL_30_DAYS,
-  makeIntervalSelectors({ val: '90d', label: '3M', interval: '8h' }),
-  makeIntervalSelectors({ val: '183d', label: '6m', interval: '1d' }),
-  makeIntervalSelectors({
-    label: 'All',
-    interval: '1d',
-    from: CRYPTO_ERA_START_DATE
-  })
-]
 
 const DashboardMetricChart = ({
   className,
